Move things closer to where they should be, for clarity
This is moving into spike territory though now
